Replaced cell elevation with a HexCellType class
The logic for setting cell elevation now no longer relies on the naively arbitrary indices of the `HexGrid.cellMaterials` list.
It's instead replaced with logic based on an array - `HexGrid.cellTypes` - containing the material and the associated elevation for each type of cell.

Also, moved some of the logic in the previous `HexGrid.ElevateCell()` to the setter of `HexCell.CellType`.