Loot rework
Added a drop chance for Loot
Loot now has rigidbodies, so they can be pushed and interacted with physically
Removed obsolote loot variations and cleaned up Loot folder structure
Added the Swedish Palette to an Editor folder, making it accessible for future use.
Revamped the look of the loot crystal, giving it a more similar look to the base crystal.
Adjusted the animation through code to account for the new transparancy effect.