Simplified weighted random enemy spawning algorithm
The previous implementation of `ChooseRandomEnemy()` had a time complexity of O(log(n)), but was unnecessarily complex when the list of enemies
will probably always be very short (currently of length 2), and it added a fairly high threshold for understanding and maintaining the code.
This implementation has a time complexity of O(n), but is a lot simpler and should be quicker to understand.