HealthBar visual and gameplay update!
N.B. This is a scene changing commit!
Added a new shadergraph for glowing ui elements.
Created a glowing healthbar with a procedural voronoi texture tiling it giving it an animated vibe.
Added a red followBar which will tick down after the base has not received damage any damage after 2 seconds.
Created a Background element to the healthbar to give the empty bar more contrast.
Edited UIManager.cs to control the hp logic
Moved UIManager.cs to the HUD instead of having it as a separate object.
Minor fix to allow InventoryManager to reference the  UIManager singelton should it not be assigned.