Added a new options menu
Includes resolution options
Added a highlighting effect on text material allowing text to be more visible.
Changed the ControllerButtonNavigator.cs to apply the colors of the original button ui elements that we are moving through.
Added the Options menu to build settings and connected it up to the rest of the game.